Revert "Merge branch 'trinodb:master' into master"
This reverts commit e33030c9cbfa68e8727b8a411cd5c46ea4f14458, reversing
changes made to 4f70f96f127c8eadc1b007a9ffac6d8bbdcc9e07.

+++ core/trino-main/src/main/java/io/trino/connector/informationschema/InformationSchemaPageSource.java
@@ -250,7 +250,7 @@ public class InformationSchemaPageSource
     {
         for (Map.Entry<SchemaTableName, List<ColumnMetadata>> entry : listTableColumns(session, metadata, accessControl, prefix).entrySet()) {
             SchemaTableName tableName = entry.getKey();
-            long ordinalPosition = 1;
+            int ordinalPosition = 1;
 
             for (ColumnMetadata column : entry.getValue()) {
                 if (column.isHidden()) {
